requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-docx==1.1.0
openai==1.12.0
python-multipart==0.0.6
pydantic==2.5.3
httpx==0.26.0
python-dotenv==1.0.0
pdf2docx==0.5.8
PyMuPDF==1.23.8
